--- backend/src/main/java/com/example/SKALA_Mini_Project_1/Seats/dto/RequestSeatsDto.java
@@ -1,5 +1,0 @@
-package com.example.SKALA_Mini_Project_1.Seats.dto;
-
-public class RequestSeatsDto {
-
-}
