--- backend/src/main/java/com/example/SKALA_Mini_Project_1/Social/social.java
@@ -1,5 +1,0 @@
-package com.example.SKALA_Mini_Project_1.Social;
-
-public class social {
-
-}
