--- backend/src/main/java/com/example/SKALA_Mini_Project_1/Users/user.java
@@ -1,5 +1,0 @@
-package com.example.SKALA_Mini_Project_1.Users;
-
-public class user {
-
-}
