+++ backend/src/main/java/com/example/SKALA_Mini_Project_1/domain/Seats/entity/Seat.java
@@ -1,0 +1,5 @@
+package com.example.SKALA_Mini_Project_1.domain.Seats.entity;
+
+public class Seat {
+
+}
