+++ backend/src/main/java/com/example/SKALA_Mini_Project_1/domain/Users/user.java
@@ -1,0 +1,5 @@
+package com.example.SKALA_Mini_Project_1.domain.Users;
+
+public class user {
+
+}
