+++ backend/src/main/java/com/example/SKALA_Mini_Project_1/domain/Waiting/waiting.java
@@ -1,0 +1,5 @@
+package com.example.SKALA_Mini_Project_1.domain.Waiting;
+
+public class waiting {
+
+}
