+++ backend/src/main/resources/application.properties
@@ -4,3 +4,6 @@ spring.datasource.username=${DB_USER}
 spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 server.port=${SERVER_PORT:8080}
+# Redis 설정
+spring.data.redis.host=localhost
+spring.data.redis.port=6379
