+++ infra/docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3.8'
+
+services:
+
+  redis:
+    image: redis:7
+    container_name: concert-redis
+    ports:
+      - "6379:6379"
+    command: redis-server --appendonly yes
